Cleans up depreciated and unused code follow up to https://github.com/RedHatInsights/insights-advisor-frontend/pull/217
