+++ src/PresentationalComponents/Charts/SummaryChart/SummaryChart.js
