+++ src/PresentationalComponents/Charts/SummaryChart/SummaryChartItem.js
