+++ src/PresentationalComponents/Charts/SummaryChart/SummaryChartItem.scss
@@ -1,5 +1,5 @@
-@import '~@red-hat-insights/insights-frontend-components/Utilities/_variables';
-@import '~@red-hat-insights/insights-frontend-components/Utilities/_mixins';
+@import '@red-hat-insights/insights-frontend-components/Utilities/_variables.scss';
+@import '@red-hat-insights/insights-frontend-components/Utilities/_mixins.scss';
 
 
 .summary-chart a{
