+++ src/PresentationalComponents/Charts/SummaryChart/SummaryChartItem.test.js
