--- src/PresentationalComponents/Skeletons/_Skeleton.scss
@@ -1,25 +1,0 @@
-@import '~@red-hat-insights/insights-frontend-components/Utilities/_variables';
-@import '~@red-hat-insights/insights-frontend-components/Utilities/_mixins';
-
-$skeleton-base-color: #f5f5f5;
-$skeleton-dark-color: #e8e8e8;
-
-.skeleton {
-    &-xs { width: 16%; }
-    &-sm { width: 33%; }
-    &-md { width: 66%; }
-    &-lg { width: 100%; }
-    animation-duration: 1s;
-    animation-iteration-count: infinite;
-    animation-name: skeletonGradient;
-    animation-timing-function: linear;
-    background: $ins-color--gray;
-    background: linear-gradient(to right, $skeleton-base-color 10%, $skeleton-dark-color 18%, $skeleton-base-color 33%);
-    @include rem('background-size', 1000px 50px);
-    @include rem('border-radius', 5px);
-}
-
-@keyframes skeletonGradient{
-    0% { @include rem('background-position', -500px 0); }
-    100%{ @include rem('background-position', 500px 0); }
-}
