--- src/SmartComponents/Actions/_actions.scss
@@ -1,5 +1,0 @@
-.page__actions {
-  .actions__view--title {
-    text-transform: capitalize;
-  }
-}
