--- src/SmartComponents/Rules/_rules.scss
@@ -1,3 +1,0 @@
-.page__rules {
-    // Filler
-}
